improv: update trimming slider ui and refactor usecases
