+++ app/src/main/kotlin/eu/peernetwork/app/module/media/UsecaseModule.kt
@@ -2,16 +2,11 @@ package eu.peernetwork.app.module.media
 
 import dagger.Module
 import dagger.Provides
-import eu.peernetwork.media.core.usecase.CroppedPreviewUseCase
-import eu.peernetwork.media.core.usecase.DecodeAndCacheThumbnailUsecase
 import eu.peernetwork.media.core.usecase.MediaEncoderUsecase
 import eu.peernetwork.media.core.usecase.MetaDataUsecase
-import eu.peernetwork.media.core.usecase.VideoTrimUsecase
-import eu.peernetwork.media.ui.usecase.CroppedPreviewUsecaseDelegate
-import eu.peernetwork.media.ui.usecase.DecodeAndCacheThumbnailUsecaseDelegate
 import eu.peernetwork.media.ui.usecase.MediaEncoderUsecaseDelegate
 import eu.peernetwork.media.ui.usecase.MetaDataUsecaseDelegate
-import eu.peernetwork.media.ui.usecase.VideoTrimUsecaseDelegate
+
 import javax.inject.Singleton
 
 @Module
@@ -24,19 +19,4 @@ object UsecaseModule {
     @Singleton
     fun bindMetaDataUsecase(delegate: MetaDataUsecaseDelegate): MetaDataUsecase = delegate
 
-    @Provides
-    @Singleton
-    fun bindVideoTrimUsecase(delegate: VideoTrimUsecaseDelegate): VideoTrimUsecase = delegate
-
-    @Provides
-    @Singleton
-    fun bindCroppedPreviewUsecase(
-        delegate: CroppedPreviewUsecaseDelegate
-    ): CroppedPreviewUseCase = delegate
-
-    @Provides
-    @Singleton
-    fun bindDecodeAndCacheThumbnail(
-        delegate: DecodeAndCacheThumbnailUsecaseDelegate
-    ): DecodeAndCacheThumbnailUsecase = delegate
 }
